Done create admin fee functionality

+++ database/migrations/2024_07_18_131313_create_admin_fees_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('admin_fees', function (Blueprint $table) {
             $table->id();
-            $table->decimal('fee', 10, 2);
+            $table->integer('fee');
             $table->string('status');
             $table->timestamps();
         });
